feat: add vue router, pinia and fix some bugs
