+++ src/main.js
@@ -1,6 +1,15 @@
-import './assets/main.css'
+import { createApp } from 'vue';
+import { createPinia } from 'pinia';
 
-import { createApp } from 'vue'
-import App from './App.vue'
+import App from './App.vue';
+import router from './router';
 
-createApp(App).mount('#app')
+import './assets/main.css';
+
+const app = createApp(App);
+const pinia = createPinia();
+
+app.use(router);
+app.use(pinia);
+
+app.mount('#app');
